requirements.txt
openai
python-dotenv
whisper
webrtcvad
pydub
sounddevice
numpy
requests
colorama
tqdm
termcolor
